+++ .idea/encodings.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/affectSimulator/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/affectSimulator/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/affectTracker/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/affectTracker/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/cobotSimulatorOne/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/cobotSimulatorOne/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/cobotSimulatorTwo/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/cobotSimulatorTwo/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/eyeSimulator/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/eyeSimulator/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/headSimulatorOne/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/headSimulatorOne/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/headSimulatorTwo/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/headSimulatorTwo/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/library/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/library/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
+  </component>
+</project>
